make deployment easier
