+++ backend/config/config.json
@@ -9,6 +9,6 @@
   },
   "production": {
     "dialect": "sqlite",
-    "storage": "database/database.sqlite"
+    "storage": "../database/database.sqlite"
   }
 }
