--- backend/scripts/drop.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-npx sequelize db:migrate:undo:all
-echo "Dropped all tables"
