--- backend/scripts/migrate.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-npx sequelize db:migrate
-echo "Migration complete."
