--- backend/scripts/seed.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-npx sequelize db:seed:all
